Implements wonder special cards on the backend.
Needs migration done - I had a problem generating it again.

Playing wonder cards for Halicarnassus will stop the game until the front end functionality for it is implemented. When an alert is needed in order to play a discarded card, a section of the JSON object 'discarded' will show up.